sentinelops-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
redis==5.0.1
celery==5.3.6
openai==1.14.3
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
numpy==1.26.4
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
websockets==12.0
pytest==8.1.0
pytest-asyncio==0.23.5
pandas==2.2.1
